Move anchor to top level
